requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cssselect==1.2.0
cycler==0.11.0
decorator==5.1.1
Django==4.2
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.12.2
Flask==2.3.2
fonttools==4.39.3
fpdf==1.7.2
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
kiwisolver==1.4.4
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.3
openai==0.27.8
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
parsel==1.8.1
Pillow==9.5.0
plotly==5.14.1
protobuf==3.20.3
psycopg2-binary==2.9.6
pyarrow==11.0.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
PyQt6==6.5.0
PyQt6-Qt6==6.5.0
PyQt6-sip==13.5.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.29.0
rich==13.3.5
selectorlib==0.16.0
six==1.16.0
smmap==5.0.0
snscrape==0.7.0.20230622
soupsieve==2.4.1
sqlparse==0.4.4
streamlit==1.22.0
tenacity==8.2.2
tk==0.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
w3lib==2.1.1
Werkzeug==2.3.3
yarl==1.9.2
zipp==3.15.0
